refactor: CLI からビジネスロジッククラスを core/ に分離
