+++ pochivision/core/__init__.py
@@ -1,4 +1,7 @@
 """coreパッケージ:画像処理パイプラインの実行・管理機能を提供します."""
 
 # flake8: noqa: F401
+from .feature_extraction import FeatureExtractionRunner
+from .fft_visualization import SimpleFFTVisualizer
 from .pipeline_executor import PipelineExecutor
+from .profile_processing import ProfileProcessor
